work/code/mcp/requirements.txt
# Core MCP Server Dependencies
websockets>=12.0
pydantic>=2.0.0

# Storage Backend
gitpython>=3.1.40
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# Vector Database
chromadb>=0.4.15
sentence-transformers>=2.2.2
numpy>=1.24.0

# AI Integration
openai>=1.0.0
anthropic>=0.8.0

# LaTeX Processing
pylatex>=1.4.1
jinja2>=3.1.0

# Utilities
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
